feat(lld/myledger): proper states for CLS + breakdown big components
