+++ apps/ledger-live-desktop/src/renderer/screens/customImage/Step1ChooseImage.tsx
@@ -8,7 +8,7 @@ import ImportNFTButton from "~/renderer/components/CustomImage/ImportNFTButton";
 import NFTGallerySelector from "~/renderer/components/CustomImage/NFTGallerySelector";
 import { StepProps } from "./types";
 import StepContainer from "./StepContainer";
-import { Flex, InfiniteLoader } from "@ledgerhq/react-ui";
+import { Flex, IconsLegacy, InfiniteLoader, Link } from "@ledgerhq/react-ui";
 import StepFooter from "./StepFooter";
 import {
   ImageLoadFromNftError,
@@ -19,6 +19,10 @@ import useIsMounted from "@ledgerhq/live-common/hooks/useIsMounted";
 import TrackPage from "~/renderer/analytics/TrackPage";
 import { analyticsPageNames, analyticsFlowName } from "./shared";
 import { useTrack } from "~/renderer/analytics/segment";
+import { setDrawer } from "~/renderer/drawers/Provider";
+import RemoveCustomImage from "../manager/DeviceDashboard/DeviceInformationSummary/RemoveCustomImage";
+import { useSelector } from "react-redux";
+import { lastSeenCustomImageSelector } from "~/renderer/reducers/settings";
 
 type Props = StepProps & {
   onResult: (res: ImageBase64Data) => void;
@@ -99,6 +103,12 @@ const StepChooseImage: React.FC<Props> = props => {
     [isMounted, onError, selectedNftId],
   );
 
+  const lastSeenCustomImage = useSelector(lastSeenCustomImageSelector);
+
+  const onRemove = useCallback(() => {
+    setDrawer(RemoveCustomImage, {});
+  }, []);
+
   return (
     <StepContainer
       footer={
@@ -148,6 +158,17 @@ const StepChooseImage: React.FC<Props> = props => {
               });
             }}
           />
+          {lastSeenCustomImage?.size ? (
+            <Link
+              size="medium"
+              color="error.c60"
+              mt={10}
+              onClick={onRemove}
+              Icon={IconsLegacy.TrashMedium}
+            >
+              {t("removeCurrentPicture.cta")}
+            </Link>
+          ) : null}
         </Flex>
       ) : (
         <NFTGallerySelector handlePickNft={handlePickNft} selectedNftId={selectedNftId} />
