+++ apps/ledger-live-desktop/src/renderer/screens/manager/DeviceDashboard/AppDepsUnInstallModal.tsx
