+++ apps/ledger-live-desktop/src/renderer/screens/manager/DeviceDashboard/AppsList/Placeholder.tsx
