+++ apps/ledger-live-desktop/src/renderer/screens/manager/DeviceDashboard/AppsList/Progress.tsx
