+++ apps/ledger-live-desktop/src/renderer/screens/manager/DeviceDashboard/AppsList/index.tsx
@@ -13,14 +13,14 @@ import Card from "~/renderer/components/Box/Card";
 import Box from "~/renderer/components/Box";
 import Text from "~/renderer/components/Text";
 import TabBar from "~/renderer/components/TabBar";
-import Item from "./Item";
+import AppRow from "./Item";
 import Filter from "./Filter";
 import Sort from "./Sort";
 import UninstallAllButton from "./UninstallAllButton";
 import { openModal } from "~/renderer/actions/modals";
 import debounce from "lodash/debounce";
 import InstallSuccessBanner from "./InstallSuccessBanner";
-import SearchBox from "../../accounts/AccountList/SearchBox";
+import SearchBox from "../../../accounts/AccountList/SearchBox";
 import { App } from "@ledgerhq/types-live";
 import { AppType, SortOptions } from "@ledgerhq/live-common/apps/filtering";
 import NoResults from "~/renderer/icons/NoResults";
@@ -128,7 +128,7 @@ const AppsList = ({
   const mapApp = useCallback(
     (app: App, appStoreView: boolean, onlyUpdate?: boolean, showActions?: boolean) => {
       return (
-        <Item
+        <AppRow
           optimisticState={optimisticState}
           state={state}
           key={`${appStoreView ? "APP_STORE" : "DEVICE_TAB"}_${app.name}`}
